Add files via upload
I made this project using solenoid valves and arduino. 
Advantages: Because batch production involves small batches, it is good for quality control. For example, if there is a mistake in the process, it can be fixed without as much loss compared to mass production. This can also save money by taking less risk for newer plans and products etc. As a result, this allows batch manufacturing to be changed or modified depending on company needs.
Disadvantages: There can be downtime between individual batches. Or if the product is constantly changing or being modified throughout the process, this also can cost downtime. Other disadvantages are that smaller batches need more planning, scheduling and control over the process and collecting data.